refactor: Class Name Change
RacingCar -> RacingCarGame, RacingCarTest -> RacingCarGameTest 로 이름 변경
