+++ src/main/java/study/step3/RacingCarApp.java
@@ -2,7 +2,7 @@ package study.step3;
 
 public class RacingCarApp {
     public static void main(String[] args) {
-        RacingCar racingCar = new RacingCar();
+        RacingCarGame racingCar = new RacingCarGame();
         racingCar.start();
     }
 }
